+++ lib/statsd/instrument/metric.rb
@@ -91,6 +91,7 @@ class StatsD::Instrument::Metric
     ms: 'measure',
     g:  'gauge',
     h:  'histogram',
+    d:  'distribution',
     kv: 'key/value',
     s:  'set',
   }